finish storing

+++ src/main/kotlin/Memory.kt
@@ -2,8 +2,8 @@ package org.example
 
 abstract class Memory {
     val memory = UByteArray(4000)
-    fun read(address: Int): UByte {
-        return memory[address]
+    fun read(address: UShort): UByte {
+        return memory[address.toInt()]
     }
     abstract fun write()
 }
